Add Kosovo
Kosovo isn't officially added to ISO 3166 yet.  According to https://en.wikipedia.org/wiki/Kosovo : XK is a "user assigned" ISO 3166 code not designated by the standard, but used by the European Commission, Switzerland, the Deutsche Bundesbank and other organisations.
Also on https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3 : UNK identifies Kosovo residents to whom travel documents were issued by the United Nations Interim Administration in Kosovo (UNMIK)
As it does not have an ISO 3166-3 numeric code, I just used 999 which is within the user assigned range, but I'm not sure if this is appropriate, or if it should just be left as an empty string.